pyproject.toml
[project]
name = "testdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.116.1",
    "fastmcp>=2.0.0",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.7",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.6.6",
    "mcp>=1.13.1",
    "numpy>=1.24.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.2",
    "streamlit>=1.49.0",
    "tavily-python>=0.7.11",
    "torch>=2.0.0",
    "transformers>=4.21.0",
    "wikipedia>=1.4.0",
    "wikipedia-api>=0.8.1",
]
